refactor: extract Program.cs service registrations into Startup/ extensions (#1144)
Splits the 1,976-line Program.cs into nine focused Startup/ files, dropping
the host composition root to ~1,088 LOC while preserving the exact
registration order in every extracted block (deploy backends, batch
backends, hosted-service activation, options binders, feature-event
plumbing, and the JsonSerializerContext resolver chain are all order-
sensitive). The middleware pipeline, host-builder construction, and
Aspire/Serilog wiring stay inline so the request-pipeline ordering remains
trivially auditable.
